feat: Add utility function for merging class names
This commit adds a new utility function, `cn`, to the `utils.ts` file. The `cn` function takes in multiple class names as arguments and merges them using the `clsx` and `twMerge` functions from the `clsx` and `tailwind-merge` libraries respectively. This utility function will be useful for combining class names in the project.

Ref: #123